Consolidate CSS color parsing into com.codename1.ui.CSSColor
The vector map engine shipped its own maps/vector/ColorParser, a weaker copy of
the color-string parsing already in CSSGradientParser (no named colors, no
#rgba, no percent components) that also hand-rolled a split() instead of using
StringUtil.tokenize. Extract the canonical parser -- the CSSGradientParser
superset, named-color table included -- into a new public CSSColor utility with
parse(String) (throws) and parse(String, int default) (lenient), splitting
rgb() components via StringUtil.tokenize.

CSSGradientParser.parseColor and the map style engine now both delegate to
CSSColor, so the grammar lives in exactly one place; delete ColorParser. The
map engine gains named-color / percent / #rgba support for free. Tests
redirected to CSSColor (behaviour-identical) plus a case for the newly
available forms.
